requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dataclasses==0.6
Django==3.2.6
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.2
